+++ src/main/java/com/gmail/netcracker/application/service/interfaces/FolderService.java
@@ -11,21 +11,21 @@ public interface FolderService {
 
     List<Folder> folderList();
 
-    Folder getFolder(int folderId);
+    Folder getFolder(Long folderId);
 
-    void delete(int folderId);
+    void delete(Long folderId);
 
     void update(Folder folder);
 
-    List<Note> getNoteListIntoFolder(int folderId);
+    List<Note> getNoteListIntoFolder(Long folderId);
 
     List<User> getFriendsToShare(List<User> friendsThatHaveAccessList);
 
-    List<User> getFriendsThatHaveAccess(int folderId);
+    List<User> getFriendsThatHaveAccess(Long folderId);
 
-    void allowAccessToFolder(int folderId, int userId);
+    void allowAccessToFolder(Long folderId, Long userId);
 
-    void disableAccessToFolder(int folderId, int friendId);
+    void disableAccessToFolder(Long folderId, Long friendId);
 
     List<Folder> sharedFoldersToMe();
 }
